requirements.txt
beautifulsoup4==4.8.0
Flask==1.0.2
python-dateutil==2.6.1
PyYAML==5.1.2
requests==2.22.0
SQLAlchemy==1.3.8
